Update 3D pohon dan quest static untuk HandleQuest

+++ Assets/Code/Script/Explore/Player/SaveLastPosisi.cs
@@ -8,7 +8,7 @@ namespace Nivandria.Explore
     public class SaveLastPosisi : MonoBehaviour
     {
         [Header("Index Scene")]
-        [SerializeField] int Scene;
+        [SerializeField] byte Scene;
         [Header("Players")]
         [SerializeField] Transform Players;
